+++ sdk/resourcemanager/appcontainers/armappcontainers/managedenvironments_client.go
@@ -519,7 +519,7 @@ func (client *ManagedEnvironmentsClient) update(ctx context.Context, resourceGro
 	if err != nil {
 		return nil, err
 	}
-	if !runtime.HasStatusCode(resp, http.StatusAccepted) {
+	if !runtime.HasStatusCode(resp, http.StatusOK, http.StatusAccepted) {
 		return nil, runtime.NewResponseError(resp)
 	}
 	return resp, nil